依旧是坑，所以就不导出apk了
做了设置界面，现在头像还没做好

+++ app/src/main/java/org/miaowo/miaowo/impl/interfaces/Users.java
@@ -15,7 +15,7 @@ public interface Users {
      * @param id 获取用户的ID
      * @return 返回的用户，若未登录则返回用户id为-1的用户(Guest)
      */
-    User getUser(long id);
+    User getUser(int id);
 
     /**
      * 根据用户名获取用户信息
